Registro de usuarios con hardcodeo
Archivos que toque:

ControladorLogin: Comente el codigo base de taller para probar login
(pero no realize ninguna funcion)
ControladorRegistroUsuario: Hardcodeo y valido x email de usuario
Usuario: cambie atributos
altaProducto: Redireccione el menu, donde dice Usuario
holaMundo: cambie la vista
home: cambie el titulo
loginUsuario: cambie titulos y redireccionamientos. Agregue boton

+++ Farmaclick/src/main/java/ar/edu/unlam/tallerweb1/modelo/Usuario.java
@@ -10,7 +10,7 @@ public class Usuario {
 	
 	@Id 
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id; /* era long */
 	private String nombre;
 	private String apellido;
 	private String email;
@@ -19,10 +19,10 @@ public class Usuario {
 
 	/*GETTERS & SETTERS*/	
 	
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getNombre() {
